+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Made `TokenOperation`s deserializable from CBOR.
   `TokenUpdate.builder().operationsSerialized()` can be used for deserialization
 - Added `TaggedTokenHolderAccount.getAddress()` returning the address in a convenient form
+- Removed checked exception from `AtomicStatement.canBeProvedBy()`
 
 ## 11.0.2